Move abstract sample test to main game project
